requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
decorator==4.4.2
Django==4.2.3
django-picklefield==3.1
django-q2==1.5.4
hachoir==3.2.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.25.1
Pillow==10.0.0
proglog==0.1.10
pyaes==1.6.1
pyasn1==0.5.0
PyMySQL==1.1.0
requests==2.31.0
rsa==4.9
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
Telethon==1.28.5
tqdm==4.65.0
urllib3==2.0.4
